requirements.txt
fastapi>=0.104.1
uvicorn>=0.34.0
openai>=1.59.6
python-dotenv>=1.0.1
pydantic>=2.10.6
pypdf>=5.1.0
PyPDF2>=3.0.1
numpy>=2.2.0
